refactor: move old sync handlers to mod v0
